Cleanup; Refactor; Move geometry utils to their own file; Update imports
