+++ web_ui/src/core/annotations/vec2.ts
@@ -83,10 +83,6 @@ export const abs = (value: Vec2): Vec2 => {
     };
 };
 
-export const toTurfPoint = ({ x, y }: Vec2): number[] => {
-    return [x, y];
-};
-
 export const getAngle = (direction: Vec2): number => {
     return Math.atan2(direction.y, direction.x);
 };
