+++ web_ui/src/pages/annotator/tools/edit-tool/edit-rotated-bounding-box/edit-rotated-bounding-box.test.tsx
@@ -78,7 +78,7 @@ describe('EditRectangleTool', (): void => {
             x: (startPoint.x - endPoint.x) / zoom,
             y: (startPoint.y - endPoint.y) / zoom,
         };
-        expect(onComplete).toBeCalledWith({
+        expect(onComplete).toHaveBeenCalledWith({
             ...annotation,
             shape: {
                 ...shape,
@@ -134,7 +134,7 @@ describe('EditRectangleTool', (): void => {
                 clientY: zoom * endPoint.y,
             });
 
-            expect(onComplete).toBeCalledWith({
+            expect(onComplete).toHaveBeenCalledWith({
                 ...annotation,
                 shape: { ...shape, ...expectedEndPoint },
             });
@@ -394,7 +394,7 @@ describe('EditRectangleTool', (): void => {
             clientY: endPoint.y,
         });
 
-        expect(onComplete).toBeCalledWith({ ...annotation, shape: expectedRect });
+        expect(onComplete).toHaveBeenCalledWith({ ...annotation, shape: expectedRect });
     });
 
     describe('Resizing to its minimum', () => {
@@ -437,7 +437,7 @@ describe('EditRectangleTool', (): void => {
                 clientY: endPoint.y,
             });
 
-            expect(onComplete).toBeCalledWith({ ...annotation, shape: expectedRect });
+            expect(onComplete).toHaveBeenCalledWith({ ...annotation, shape: expectedRect });
         });
     });
 });
