+++ web_ui/src/pages/annotator/tools/segment-anything-tool/segment-anything-tool.component.tsx
@@ -15,9 +15,10 @@ import { ToolType } from '../../core/annotation-tool-context.interface';
 import { useAnnotationToolContext } from '../../providers/annotation-tool-provider/annotation-tool-provider.component';
 import { useROI } from '../../providers/region-of-interest-provider/region-of-interest-provider.component';
 import { useZoom } from '../../zoom/zoom-provider.component';
+import { removeOffLimitPoints } from '../geometry-utils';
 import { SvgToolCanvas } from '../svg-tool-canvas.component';
 import { PointerType } from '../tools.interface';
-import { removeOffLimitPoints, SELECT_ANNOTATION_STYLES } from '../utils';
+import { SELECT_ANNOTATION_STYLES } from '../utils';
 import { InteractiveSegmentationPoint } from './interactive-segmentation-point.component';
 import { useSegmentAnything } from './segment-anything-state-provider.component';
 import { InteractiveAnnotationPoint } from './segment-anything.interface';
