+++ web_ui/src/pages/annotator/tools/selecting-tool/components/brush-tool.test.tsx
@@ -4,7 +4,6 @@
 import '@wessberg/pointer-events';
 
 import { fireEvent, screen, waitForElementToBeRemoved } from '@testing-library/react';
-import { area } from '@turf/turf';
 
 import { Annotation } from '../../../../../core/annotations/annotation.interface';
 import { Polygon } from '../../../../../core/annotations/shapes.interface';
@@ -17,7 +16,7 @@ import { getMockedImage, getMockedROI } from '../../../../../test-utils/utils';
 import { ProjectProvider } from '../../../../project-details/providers/project-provider/project-provider.component';
 import { AnnotationToolContext } from '../../../core/annotation-tool-context.interface';
 import { TaskProvider } from '../../../providers/task-provider/task-provider.component';
-import { shapeToTurfPolygon } from '../../utils';
+import { calculatePolygonArea, shapeToTurfPolygon } from '../../geometry-utils';
 import { BrushTool } from './brush-tool.component';
 
 const mockROI = getMockedROI();
@@ -123,7 +122,9 @@ describe('BrushTool', () => {
 
         const newPolygon = (spyUpdateAnnotation.mock.calls[0][0] as Annotation).shape;
 
-        expect(area(shapeToTurfPolygon(newPolygon))).toBeGreaterThan(area(shapeToTurfPolygon(testPolygon)));
+        expect(calculatePolygonArea(shapeToTurfPolygon(newPolygon))).toBeGreaterThan(
+            calculatePolygonArea(shapeToTurfPolygon(testPolygon))
+        );
     });
 
     it('shape decreases its size when the user clicks outside and touches the polygon', async () => {
@@ -152,6 +153,8 @@ describe('BrushTool', () => {
 
         const newPolygon = (spyUpdateAnnotation.mock.calls[0][0] as Annotation).shape;
 
-        expect(area(shapeToTurfPolygon(testPolygon))).toBeGreaterThan(area(shapeToTurfPolygon(newPolygon)));
+        expect(calculatePolygonArea(shapeToTurfPolygon(testPolygon))).toBeGreaterThan(
+            calculatePolygonArea(shapeToTurfPolygon(newPolygon))
+        );
     });
 });
