+++ web_ui/src/pages/annotator/tools/watershed-tool/watershed-tool.component.tsx
@@ -11,10 +11,11 @@ import { ToolType } from '../../core/annotation-tool-context.interface';
 import { useAddUnfinishedShape } from '../../hooks/use-add-unfinished-shape.hook';
 import { useROI } from '../../providers/region-of-interest-provider/region-of-interest-provider.component';
 import { useZoom } from '../../zoom/zoom-provider.component';
+import { isPolygonValid } from '../geometry-utils';
 import { MarkerTool } from '../marker-tool/marker-tool.component';
 import { Marker } from '../marker-tool/marker-tool.interface';
 import { ToolAnnotationContextProps } from '../tools.interface';
-import { drawingStyles, isPolygonValid } from '../utils';
+import { drawingStyles } from '../utils';
 import { BrushSizeCursor } from './brush-size-cursor.component';
 import { BACKGROUND_LABEL_MARKER_ID, formatAndAddAnnotations, getScaleValue } from './utils';
 import { useWatershedState } from './watershed-state-provider.component';
