+++ web_ui/tests/features/annotator/tools/ritm-tool.spec.ts
@@ -6,7 +6,7 @@ import { expect, Page } from '@playwright/test';
 import { Point, Rect } from '../../../../src/core/annotations/shapes.interface';
 import { ShapeType } from '../../../../src/core/annotations/shapetype.enum';
 import { paths } from '../../../../src/core/services/routes';
-import { isShapeWithinRoi } from '../../../../src/pages/annotator/tools/utils';
+import { isShapeWithinRoi } from '../../../../src/pages/annotator/tools/geometry-utils';
 import { annotatorTest as test } from '../../../fixtures/annotator-test';
 import { OpenApiFixtures } from '../../../fixtures/open-api';
 import { annotatorUrl, userAnnotationsResponse } from '../../../mocks/classification/mocks';
